+++ CarService/src/test/java/org/example/service/OrderServiceTest.java
@@ -72,7 +72,7 @@ public class OrderServiceTest {
 
     @Test
     void removeOrderNotExistingEntityTest() {
-        assertThatThrownBy(() -> orderService.removeOrder(100500)).isInstanceOf(OrderNotFoundException.class);
+        assertThatThrownBy(() -> orderService.removeOrder(100500L)).isInstanceOf(OrderNotFoundException.class);
     }
 
     @Test
@@ -80,7 +80,7 @@ public class OrderServiceTest {
         repairerService.add(REPAIRER4);
         orderService.assignRepairer(ORDER3, REPAIRER4.getId());
 
-        assertThat(orderService.getOrders().get((ORDER3.getId() - 1)).getRepairers()).contains(REPAIRER4);
+        assertThat(orderService.getOrders().get((int)(ORDER3.getId() - 1)).getRepairers()).contains(REPAIRER4);
     }
 
     @Test
@@ -89,7 +89,7 @@ public class OrderServiceTest {
         garageService.add(GARAGE_SLOT);
         orderService.assignGarageSlot(ORDER3, GARAGE_SLOT.getId());
 
-        assertThat(orderService.getOrders().get((ORDER3.getId() - 1)).getGarageSlot()).isEqualTo(GARAGE_SLOT);
+        assertThat(orderService.getOrders().get((int) (ORDER3.getId() - 1)).getGarageSlot()).isEqualTo(GARAGE_SLOT);
     }
 
     @Test
